// pages/api/addCustomerOrders/order.ts
export interface OrderTransaction {
    AutoID?: number;
    TransactionDateTime: string | Date;
    TransactionID: number;
    TransactionKey: string;
    OrderKey: string;
    OrderID: number;
    OrderDateTime: string | Date;
    StationID?: number;
    EmployeeID?: number;
    RevenueCenterTypeID?: number;
    MenuItemID?: number;
    MenuItemKey?: string;
    MenuItemText?: string;
    MenuItemGroupText?: string;
    MenumItemCategoryText?: string;
    MenuItemUnitPrice?: number;
    MenuItemCost?: number;
    Quantity?: number;
    ExtendedPrice?: number;
    DiscountID?: number;
    DiscountKey?: string;
    DiscountLineAmount?: number;
    DiscountCashAmount?: number;
    DiscountTotalAmount?: number;
    TransactionStatus?: number;
    NotificationStatus?: number;
    AdditionLinePrinted?: boolean;
    TaxPercent?: number;
    RoundID?: number;
    Mod1ID?: number;
    Mod1Cost?: number;
    Mod2ID?: number;
    Mod2Cost?: number;
    Mod3ID?: number;
    Mod3Cost?: number;
    Mod4ID?: number;
    Mod4Cost?: number;
    Mod5ID?: number;
    Mod5Cost?: number;
    Mod6ID?: number;
    Mod6Cost?: number;
    Mod7ID?: number;
    Mod7Cost?: number;
    Mod8ID?: number;
    Mod8Cost?: number;
    Mod9ID?: number;
    Mod9Cost?: number;
    Mod10ID?: number;
    Mod10Cost?: number;
    Mod11ID?: number;
    Mod11Cost?: number;
    Mod12ID?: number;
    Mod12Cost?: number;
    Mod13ID?: number;
    Mod13Cost?: number;
    Mod14ID?: number;
    Mod14Cost?: number;
    Mod15ID?: number;
    Mod15Cost?: number;
    Mod16ID?: number;
    Mod16Cost?: number;
    Mod17ID?: number;
    Mod17Cost?: number;
    Mod18ID?: number;
    Mod18Cost?: number;
    Mod19ID?: number;
    Mod19Cost?: number;
    Mod20ID?: number;
    Mod20Cost?: number;
    SeatNumber?: number;
    OnHoldUntilTime?: string | Date;
    Notes?: string;
    SaleTaxAmount?: number;
    UsedPrinterID1?: number;
    UsedPrinterID2?: number;
    UsedPrinterID3?: number;
    UsedPrinterID4?: number;
    UsedPrinterID5?: number;
    UsedPrinterComplated1?: boolean;
    UsedPrinterComplated2?: boolean;
    UsedPrinterComplated3?: boolean;
    UsedPrinterComplated4?: boolean;
    UsedPrinterComplated5?: boolean;
    LineDeleted?: boolean;
    DeleteReason?: string;
    CustomField1?: string;
    CustomField2?: string;
    CustomField3?: string;
    CustomField4?: string;
    CustomField5?: string;
    EditKey: string;
    SyncKey: string;
    BranchID?: number;
    AddUserID?: number;
    AddDateTime: string | Date;
    EditUserID?: number;
    EditDateTime: string | Date;
    AddUserName?: string;
    EditUserName?: string;
    EmployeeName?: string;
    DayEnd?: boolean;
    IsProduced?: boolean;
    SplitDetail?: string;
    ExternalOrderStatus?: number;
    PromotionKey?: string;
    DiscountAmountValue?: number;
    DiscountBasisValue?: number;
    TaxPercentReduction?: number;
    UsedDiscountName?: string;
    SendOK?: boolean;
    PromotionName?: string;
    RetailData?: string;
    OrderByWeight?: boolean;
    ProductKey?: string;
    DiscountUserName?: string;
    ImportDate?: string | Date;
    OwnerKey?: string;
    AccountingCode?: string;
    MainMenuItemText?: string;
    MainMenuItemTransactionKey?: string;
    MenuItemCostDate?: string | Date;
    PosVersion?: string;
    ModuleVersion?: string;
    LabelPrinted?: boolean;
    InvoicePrinted?: boolean;
    IsMainCombo?: boolean;
    MaxQuantity?: number;
    IntegrationApprove?: boolean;
    IntegrationReferenceNo?: string;
    IntegrationCardNumber?: string;
    PriceWeightPercent?: number;
    ECommerceCode?: string;
    DiscountText?: string;
    MainMenuItemAccountingCode?: string;
    IntegrationTransactionKey?: string;
    IntegrationProductName?: string;
    DiscountDescription?: string;
    UsedCouponCode?: string;
    FixName?: string;
    FixCount?: number;
    IntegrationProductKey?: string;
    IntegrationModuleTypeID?: number;
}

// bonus_orderpayments tablosu arayüzü
export interface OrderPayment {
    AutoID?: number;
    OrderPaymentID: number;
    PaymentKey: string;
    OrderID: number;
    OrderKey: string;
    StationID?: number;
    CustomerID?: number;
    CustomerEmployeeID?: number;
    CouponNumber?: string;
    RegisterSessionID?: number;
    RevenueCenterTypeID?: number;
    PaymentDateTime: string | Date;
    EmployeeID?: number;
    RegisterNo?: number;
    PaymentMethodID?: number;
    AmountTendered?: number;
    AmountPaid?: number;
    AmountChange?: number;
    ExhangeRate?: number;
    RoundingAmount?: number;
    CurrencyID?: number;
    IsAccountPayment?: boolean;
    IsAccountSale?: boolean;
    PaymentMethodCode?: string;
    PaymentNotes?: string;
    LineDeleted?: boolean;
    DeleteReason?: string;
    CustomField1?: string;
    CustomField2?: string;
    CustomField3?: string;
    CustomField4?: string;
    CustomField5?: string;
    EditKey: string;
    SyncKey: string;
    BranchID?: number;
    AddUserID?: number;
    AddDateTime: string | Date;
    EditUserID?: number;
    EditDateTime: string | Date;
    CustomerEmployeeKey?: string;
    CustomerKey?: string;
    EmployeeKey?: string;
    PaymentMethodKey?: string;
    AddUserName?: string;
    EditUserName?: string;
    EmployeeName?: string;
    PaymentMethodName?: string;
    DayEnd?: boolean;
    SendOK?: boolean;
    RetailData?: string;
    ImportDate?: string | Date;
    GlobalBankCode?: number;
    RegisterSessionKey?: string;
    GlobalBankName?: string;
    AccountingCode?: string;
    BankAccountingCode?: string;
    IncomeAccountingCode?: string;
    PaymentStatus?: number;
    PosVersion?: string;
    ModuleVersion?: string;
    RefundDetail?: string;
    TsmUsed?: boolean;
    ingenico?: string;
    IntegrationApprove?: boolean;
    IntegrationReferenceNo?: string;
    IntegrationCardNumber?: string;
    Pan?: string;
    profilo?: string;
    FixName?: string;
    FixCount?: number;
}

// İstek arayüzü
export interface OrderRequest {
    apiKey: string;
    orderHeader: OrderHeader;
    orderTransactions: OrderTransaction[];
    orderPayments: OrderPayment[];
}